fixed price bug

+++ expenses/models.py
@@ -7,7 +7,7 @@ class Expense(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     title = models.CharField(max_length=200)
     body = models.TextField()
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=13, decimal_places=2)
     date = models.DateField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
